refactor(tree-view-item): 🔄 Moved TreeItem types into connect/types
