+++ src/connect/types/tree-item.ts
@@ -1,0 +1,33 @@
+import { ConnectDropdownMenuItem } from '@/connect/components/dropdown-menu';
+
+export enum ItemType {
+    Folder = 'folder',
+    File = 'file',
+    LocalDrive = 'local-drive',
+    NetworkDrive = 'network-drive',
+    PublicDrive = 'public-drive',
+}
+
+export enum ActionType {
+    Update = 'update',
+    New = 'new',
+}
+
+export enum ItemStatus {
+    Available = 'available',
+    AvailableOffline = 'available-offline',
+    Syncing = 'syncing',
+    Offline = 'offline',
+}
+
+export interface TreeItem<T extends string = string> {
+    id: string;
+    label: string;
+    type: ItemType;
+    action?: ActionType;
+    status?: ItemStatus;
+    expanded?: boolean;
+    children?: TreeItem<T>[];
+    isSelected?: boolean;
+    options?: ConnectDropdownMenuItem<T>[];
+}
